Bug 4104: Disabled tag selection when another tag gets edited and improved/fixed UI.

+++ java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/raceboard/tagging/TagCell.java
@@ -186,21 +186,23 @@ public class TagCell extends AbstractCell<TagDTO> {
         SafeHtml headingButtons = SafeHtmlUtils.EMPTY_SAFE_HTML;
         // preview cells do not show buttons
         if (!isPreviewCell) {
-            // create buttons with icons
-            SafeHtml deleteButton = tagCellTemplate.button(style.tagActionButton() + " " + style.tagDeleteButton(),
-                    tagCellTemplate.icon(resources.deleteIcon().getSafeUri()));
-            SafeHtml editButton = tagCellTemplate.button(style.tagActionButton() + " " + style.tagEditButton(),
-                    tagCellTemplate.icon(resources.editIcon().getSafeUri()));
-
             // TODO: As soon as permission-vertical branch got merged into master, apply
             // new permission system at this if-statement and remove this old way of
             // checking for permissions. (see bug 4104, comment 9)
             // functionality: Check if user has the permission to delete or edit this tag.
-            if (tag.getUsername().equals(userService.getCurrentUser().getName())) {
-                headingButtons = tagCellTemplate.headerButtonsDeletableAndModifyable(style.tagCellHeadingButtons(),
-                        editButton, deleteButton);
-            } else if (userService.getCurrentUser().hasRole("admin")) {
-                headingButtons = tagCellTemplate.headerButtonsDeletable(style.tagCellHeadingButtons(), deleteButton);
+            if (userService.getCurrentUser() != null) {
+                // create buttons with icons
+                SafeHtml deleteButton = tagCellTemplate.button(style.tagActionButton() + " " + style.tagDeleteButton(),
+                        tagCellTemplate.icon(resources.deleteIcon().getSafeUri()));
+                SafeHtml editButton = tagCellTemplate.button(style.tagActionButton() + " " + style.tagEditButton(),
+                        tagCellTemplate.icon(resources.editIcon().getSafeUri()));
+                if (tag.getUsername().equals(userService.getCurrentUser().getName())) {
+                    headingButtons = tagCellTemplate.headerButtonsDeletableAndModifyable(style.tagCellHeadingButtons(),
+                            editButton, deleteButton);
+                } else if (userService.getCurrentUser().hasRole("admin")) {
+                    headingButtons = tagCellTemplate.headerButtonsDeletable(style.tagCellHeadingButtons(),
+                            deleteButton);
+                }
             }
         }
